+++ src/npc/dialog.py
@@ -1,6 +1,6 @@
-from .support import resource_path
-from .settings import CHARS_PER_LINE, TB_SIZE, SCREEN_WIDTH, SCREEN_HEIGHT, TB_LAYER
-from .timer import Timer
+from ..support import resource_path
+from ..settings import CHARS_PER_LINE, TB_SIZE, SCREEN_WIDTH, SCREEN_HEIGHT, TB_LAYER
+from ..timer import Timer
 import pygame
 import json
 from jsmin import jsmin  # JSON minifier function (removes comments, notably)
